requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-openai
langchain-ollama
ollama
python-multipart
pypdf
faiss-cpu
requests
gunicorn
langchain-huggingface
gunicorn
torch
# faiss-gpu